requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
openai
numpy
transformers